mobile_automation_framework/requirements.txt
appium-python-client
pytest
pytest-html
pytest-xdist
pytest-rerunfailures
selenium
faker
python-dotenv
